Test for build_scripts, if it is already there do not download. Tweak info messages about version.xml.

+++ gluex_install.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
 mkdir -p gluex_top
 pushd gluex_top
-rm -rf build_scripts-latest latest.tar.gz build_scripts
-wget -O latest.tar.gz --no-check-certificate https://github.com/jeffersonlab/build_scripts/archive/latest.tar.gz
-tar zxf latest.tar.gz
-ln -s build_scripts-latest build_scripts
+if [ -e build_scripts ]
+    then
+    echo build_scripts already here, skip download
+else
+    echo downloading build_scripts tar file
+    rm -rf build_scripts-latest latest.tar.gz build_scripts
+    wget -O latest.tar.gz --no-check-certificate https://github.com/jeffersonlab/build_scripts/archive/latest.tar.gz
+    tar zxf latest.tar.gz
+    ln -s build_scripts-latest build_scripts
+fi
 pwd_string=`pwd`
 export GLUEX_TOP=$pwd_string
 export BUILD_SCRIPTS=$GLUEX_TOP/build_scripts
@@ -18,9 +24,9 @@ echo setenv BUILD_SCRIPTS \$GLUEX_TOP/build_scripts >> setup.csh
 echo source \$BUILD_SCRIPTS/gluex_env_version.csh $pwd_string/version.xml >> setup.csh
 if [ -f version.xml ]
     then
-    echo version.xml exists
+    echo version.xml exists, skip download
 else
-    echo getting version.xml from halldweb
+    echo getting version.xml from halldweb.jlab.org
     wget --no-check-certificate https://halldweb.jlab.org/dist/version.xml
 fi
 source $BUILD_SCRIPTS/gluex_env_version.sh
